Implement a tab which lets users filter league matches by their round
This feature uses the new API endpoint
`GET /api/competitions/{id}/league/rounds/{roundNumber}` to
display matches based on the round number selected in the picker.

Additionally, the round picker is "smart" and "knows" how many matches
had already been played in the competition, so that it's set to
display the matches from the latest round by default
(i.e., after 5 rounds, the picker's default round is set to 5).